Download both the release artifacts in extended mode for LLVM >= 17
The Espressif-provided LLVM toolchain no longer has the libs-only
tarball as a strict subset of the full tarball.  Instead the "full"
tarball now contains a clang with libclang built in.  To work around
this we now need to get _both_ tarballs if we're in extended LLVM mode.

Cf. https://github.com/espressif/llvm-project/issues/99

Signed-off-by: Johannes Löthberg <johannes.loethberg@elokon.com>